// src/app/models/recipients.model.ts
export class RecipientsModel {
  Id: number;
  EntityType: string;
  EntityName: string;
  recipients: string[];
  EntityId: string[];
  NotificationId: number;
  CreatedById: number;
  UpdatedById: number;
  CreatedByRoleId: number;
  UpdatedByRoleId: number;
  WFId: number;
  WFLabel: string;
  ProjectId: number;
  StepId: number;
}